refactor: pageコンポーネントのディレクトリ名をpagesに変更
