+++ src/components/pages/Top/index.stories.tsx
@@ -1,9 +1,9 @@
 import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
-import { TopPage } from 'src/components/page/Top/index';
+import { TopPage } from 'src/components/pages/Top/index';
 
 export default {
   component: TopPage,
-  title: 'page/Top',
+  title: 'pages/Top',
 } as ComponentMeta<typeof TopPage>;
 
 export const General: ComponentStoryObj<typeof TopPage> = {};
